Work on Creating/Deleting move and error handling

+++ MovieLand/MovieLand.Services/Printer/PrinterService.cs
@@ -33,7 +33,7 @@ namespace MovieLand.Services.Printer
        public void ShowAvailableCommands()
         {
             this.SetColorToGreen();
-            var result = "Available commands: help; info [id]; ";
+            var result = "Available commands: help; info [id]; list [number]; ";
 
             if (this.userService.IsUserAuthenticated())
             {
@@ -78,7 +78,8 @@ namespace MovieLand.Services.Printer
                 .AppendLine("watched [id] -> by giving the id of the movie, the LOGGED user adds it to their collection of watched movies")
                 .AppendLine("review [id] -> after giving the id of a movie, the user can submit a review for the movie they have watched")
                 .AppendLine("info [id] -> the user is shown all the information about a movie - title, plot, etc")
-                .AppendLine("create/delete [id] -> only the admin has access to these commands which can help him moderate the platform");
+                .AppendLine("create/delete [id] -> only the admin has access to these commands which can help him moderate the platform")
+                .AppendLine("list [number] -> shows the page with that number");
 
             Console.WriteLine(sb.ToString().Trim());
             this.ClearConsoleColor();
@@ -86,58 +87,65 @@ namespace MovieLand.Services.Printer
 
         public void StartingCreatingMovie()
         {
-            if (this.userService.IsUserAdmin())
+            if (this.userService.IsUserAuthenticated() && this.userService.IsUserAdmin())
             {
                 Console.Clear();
-                Console.WriteLine("Movie id:");
-                int id = int.Parse(Console.ReadLine());
-                Console.WriteLine("Movie title:");
+                Console.Write("Movie title: ");
                 string title = Console.ReadLine();
-                Console.WriteLine("Movie plot:");
+                Console.Write("Movie plot: ");
                 string plot = Console.ReadLine();
-                Console.WriteLine("Movie producer:");
+                Console.Write("Movie producer: ");
                 string producer = Console.ReadLine();
-                Console.WriteLine("Movie genre:");
+                Console.Write("Movie genre: ");
                 string genre = Console.ReadLine();
-                Console.WriteLine("Movie actors:");
+                Console.Write("Movie actors: ");
                 string actors = Console.ReadLine();
-                Console.WriteLine("Movie key words:");
+                Console.Write("Movie key words: ");
                 ICollection<string> keyWords = Console.ReadLine().Split();
 
-                this.movieService.CreateMovie(id, title, plot, producer, genre, actors, keyWords);
-                Console.WriteLine("Movie was created successfully.");
+                this.movieService.CreateMovie(title, plot, producer, genre, actors, keyWords);
                 Console.Clear();
+                Console.WriteLine("Movie was created successfully.");
+            }
+            else
+            {
+                throw new Exception("Access denied!");
             }
         }
 
         public void DeleteMovie(List<string> tokens)
         {
-            int id = int.Parse(tokens[1]);
-            if (this.userService.IsUserAdmin())
-            {
-                var movie = movieService.GetMovie(id);
-                AssuranceForDeletingMovie();
-                string answer = Console.ReadLine();
-                bool isConferned = true;
-                if (answer == "Yes")
-                {
-                    isConferned = true;
-                }
-                else
-                {
-                    isConferned = false;
-                }
-                if (isConferned == true)
-                {
-                    this.movieService.DeleteMovie(id);
-                }
+            this.SetColorToRed();
 
+            int id = int.Parse(tokens[1]);
+            if (this.userService.IsUserAuthenticated() && this.userService.IsUserAdmin())
+            {
+                if (AssuranceForDeletingMovie())
+                {
+                    var isDeleted = this.movieService.DeleteMovie(id);
+
+                    Console.WriteLine(isDeleted ? "Movie deleted successfully!" : "Movie not deleted!");
+                    this.ClearConsoleColor();
+                }
+            }
+            else
+            {
+                throw new Exception("Access denied!");
             }
         }
 
-        public void AssuranceForDeletingMovie()
+        public void PrintError(string errorMessage)
         {
-            Console.WriteLine("Are you sure you want to delete this movie?");
+            this.SetColorToRed();
+            Console.WriteLine(errorMessage);
+            this.ClearConsoleColor();
+        }
+
+        private bool AssuranceForDeletingMovie()
+        {
+            Console.Write("Are you sure you want to delete this movie? [Y/n] ");
+
+            return Console.ReadLine() == "Y";
         }
 
         public string PrintHomeRow(string[] columns)
@@ -192,6 +200,11 @@ namespace MovieLand.Services.Printer
         private void SetColorToYellow()
         {
             Console.ForegroundColor = ConsoleColor.Yellow;
+        }
+
+        private void SetColorToRed()
+        {
+            Console.ForegroundColor = ConsoleColor.Red;
         }
     }
 }
